feat: cleanup types
